@@ -25,7 +25,7 @@ import uuid
 from concurrent.futures import Future, ThreadPoolExecutor
 from enum import Enum, auto
 from pathlib import Path
-from typing import Any, Optional, Union, cast
+from typing import Any, cast
 
 from flwr_datasets.common.version import package_name, package_version
 
@@ -55,7 +55,7 @@ def _configure_logger(log_level: int) -> None:
 _configure_logger(LOGGER_LEVEL)
 
 
-def log(msg: Union[str, Exception]) -> None:
+def log(msg: str | Exception) -> None:
     """Log message using logger at DEBUG level."""
     logging.getLogger(LOGGER_NAME).log(LOGGER_LEVEL, msg)
 
@@ -127,7 +127,7 @@ class EventType(str, Enum):
 
 # Use the ThreadPoolExecutor with max_workers=1 to have a queue
 # and also ensure that telemetry calls are not blocking.
-state: dict[str, Union[Optional[str], Optional[ThreadPoolExecutor]]] = {
+state: dict[str, str | None | ThreadPoolExecutor | None] = {
     # Will be assigned ThreadPoolExecutor(max_workers=1)
     # in event() the first time it's required
     "executor": None,
@@ -143,7 +143,7 @@ state: dict[str, Union[Optional[str], Optional[ThreadPoolExecutor]]] = {
 # pylint: disable-next=unsubscriptable-object
 def event(
     event_type: EventType,
-    event_details: Optional[dict[str, Any]] = None,
+    event_details: dict[str, Any] | None = None,
 ) -> Future:  # type: ignore
     """Submit create_event to ThreadPoolExecutor to avoid blocking."""
     if state["executor"] is None:
@@ -155,7 +155,7 @@ def event(
     return result
 
 
-def create_event(event_type: EventType, event_details: Optional[dict[str, Any]]) -> str:
+def create_event(event_type: EventType, event_details: dict[str, Any] | None) -> str:
     """Create telemetry event."""
     if state["source"] is None:
         state["source"] = _get_source_id()
