@@ -17,7 +17,6 @@
 
 
 import unittest
-from typing import Union
 from unittest.mock import Mock, patch
 
 import numpy as np
@@ -385,7 +384,7 @@ class PartitionersSpecificationForFederatedDatasets(unittest.TestCase):
         """Test if partitioners are passed directly (no recreation)."""
         num_train_partitions = 100
         num_test_partitions = 100
-        partitioners: dict[str, Union[Partitioner, int]] = {
+        partitioners: dict[str, Partitioner | int] = {
             "train": IidPartitioner(num_partitions=num_train_partitions),
             "test": IidPartitioner(num_partitions=num_test_partitions),
         }
@@ -419,7 +418,7 @@ class PartitionersSpecificationForFederatedDatasets(unittest.TestCase):
         """Test if an instantiated partitioner is passed directly."""
         num_train_partitions = 100
         num_test_partitions = 100
-        partitioners: dict[str, Union[Partitioner, int]] = {
+        partitioners: dict[str, Partitioner | int] = {
             "train": IidPartitioner(num_partitions=num_train_partitions),
             "test": num_test_partitions,
         }
@@ -433,7 +432,7 @@ class PartitionersSpecificationForFederatedDatasets(unittest.TestCase):
         """Test if an IidPartitioner partitioner is created."""
         num_train_partitions = 100
         num_test_partitions = 100
-        partitioners: dict[str, Union[Partitioner, int]] = {
+        partitioners: dict[str, Partitioner | int] = {
             "train": IidPartitioner(num_partitions=num_train_partitions),
             "test": num_test_partitions,
         }
@@ -565,7 +564,7 @@ class IncorrectUsageFederatedDatasets(unittest.TestCase):
     def test_incorrect_two_partitioners(self) -> None:
         """Test if the method raises ValueError with incorrect partitioners."""
         partitioner = IidPartitioner(num_partitions=10)
-        partitioners: dict[str, Union[Partitioner, int]] = {
+        partitioners: dict[str, Partitioner | int] = {
             "train": partitioner,
             "test": partitioner,
         }
@@ -586,7 +585,7 @@ class IncorrectUsageFederatedDatasets(unittest.TestCase):
     def test_incorrect_three_partitioners(self) -> None:
         """Test if the method raises ValueError with incorrect partitioners."""
         partitioner = IidPartitioner(num_partitions=10)
-        partitioners: dict[str, Union[int, Partitioner]] = {
+        partitioners: dict[str, int | Partitioner] = {
             "train1": partitioner,
             "train2": 10,
             "test": partitioner,
@@ -621,7 +620,7 @@ def datasets_are_equal(ds1: Dataset, ds2: Dataset) -> bool:
         return False
 
     # Iterate over each row and check for equality
-    for row1, row2 in zip(ds1, ds2):
+    for row1, row2 in zip(ds1, ds2, strict=False):
         # Ensure all keys are the same in both rows
         if set(row1.keys()) != set(row2.keys()):
             return False
