@@ -15,7 +15,7 @@
 """VerticalEvenPartitioner class."""
 # noqa: E501
 # pylint: disable=C0301, R0902, R0913
-from typing import Literal, Optional, Union
+from typing import Literal
 
 import numpy as np
 
@@ -84,21 +84,21 @@ class VerticalEvenPartitioner(Partitioner):
     def __init__(  # pylint: disable=R0917
         self,
         num_partitions: int,
-        active_party_columns: Optional[Union[str, list[str]]] = None,
-        active_party_columns_mode: Union[
+        active_party_columns: str | list[str] | None = None,
+        active_party_columns_mode: (
             Literal[
                 "add_to_first",
                 "add_to_last",
                 "create_as_first",
                 "create_as_last",
                 "add_to_all",
-            ],
-            int,
-        ] = "add_to_last",
-        drop_columns: Optional[Union[str, list[str]]] = None,
-        shared_columns: Optional[Union[str, list[str]]] = None,
+            ]
+            | int
+        ) = "add_to_last",
+        drop_columns: str | list[str] | None = None,
+        shared_columns: str | list[str] | None = None,
         shuffle: bool = True,
-        seed: Optional[int] = 42,
+        seed: int | None = 42,
     ) -> None:
         super().__init__()
 
@@ -113,7 +113,7 @@ class VerticalEvenPartitioner(Partitioner):
         self._seed = seed
         self._rng = np.random.default_rng(seed=self._seed)
 
-        self._partition_columns: Optional[list[list[str]]] = None
+        self._partition_columns: list[list[str]] | None = None
         self._partitions_determined = False
 
         self._validate_parameters_in_init()
@@ -204,7 +204,7 @@ class VerticalEvenPartitioner(Partitioner):
         self,
         all_columns: list[str],
         shared_columns: list[str],
-        active_party_column: Union[str, list[str]],
+        active_party_column: str | list[str],
     ) -> None:
         if isinstance(active_party_column, str):
             active_party_column = [active_party_column]
